Refactor ISBN-registry publisher + publisher request list and item views
