+++ src/frontend/views/isbn-registry/requests/publishers/IsbnPublisherRequestList.jsx
@@ -25,9 +25,11 @@
  *
  */
 
-import React, {useEffect, useReducer} from 'react';
-import PropTypes from 'prop-types';
+import React, {useEffect, useMemo, useReducer} from 'react';
 import {FormattedMessage} from 'react-intl';
+
+import {useHistory} from 'react-router-dom';
+import {useAuth} from 'react-oidc-context';
 
 import useSearch from '/src/frontend/hooks/useSearch';
 
@@ -36,12 +38,15 @@ import moment from 'moment';
 
 import '/src/frontend/css/common.css';
 
-import Spinner from '/src/frontend/components/common/Spinner.jsx';
 import TableComponent from '/src/frontend/components/common/TableComponent.jsx';
+import TableResultWrapper from '/src/frontend/components/common/TableResultWrapper.jsx';
+
 import SearchComponent from '/src/frontend/components/common/SearchComponent.jsx';
 
-function IsbnPublisherRequestList(props) {
-  const {authenticationToken, history} = props;
+function IsbnPublisherRequestList() {
+  const history = useHistory();
+  const {user: {access_token: authenticationToken}} = useAuth();
+
   // Component state
   const initialSearchBody = history.location?.state?.searchBody ?? {
     searchText: '',
@@ -70,6 +75,9 @@ function IsbnPublisherRequestList(props) {
     requireAuth: true,
     authenticationToken
   });
+
+  const formattedData = useMemo(() => data.results.map(formatSearchResult), [data]);
+  const hasData = formattedData && formattedData.length > 0;
 
   function updateRowsPerPage(rowsPerPage) {
     updateSearchBody({limit: rowsPerPage, offset: 0});
@@ -112,46 +120,6 @@ function IsbnPublisherRequestList(props) {
     {id: 'additionalInfo', intlId: 'form.common.additionalDetails'}
   ];
 
-  const dataComponent = setDataComponent();
-
-  function setDataComponent() {
-    if (error) {
-      return (
-        <Typography variant="h2" className="normalTitle">
-          <FormattedMessage id="errorPage.message.defaultError" />
-        </Typography>
-      );
-    }
-
-    if (loading) {
-      return <Spinner />;
-    }
-
-    if (data.totalDoc === 0) {
-      return (
-        <p>
-          <FormattedMessage id="common.noData" />
-        </p>
-      );
-    }
-
-    return (
-      <TableComponent
-        pagination
-        data={data.results.map(formatSearchResult)}
-        handleTableRowClick={handleTableRowClick}
-        headRows={headRows}
-        page={searchBody.offset !== 0 ? searchBody.offset / searchBody.limit : 0}
-        setPage={updatePageNumber}
-        totalDoc={data.totalDoc}
-        rowsPerPage={searchBody.limit}
-        setRowsPerPage={updateRowsPerPage}
-        // Those are not displayed on small screens (mobile devices etc)
-        unprioritizedRows={['email', 'additionalInfo']}
-      />
-    );
-  }
-
   return (
     <div className="listSearch">
       <Typography variant="h5">
@@ -161,14 +129,24 @@ function IsbnPublisherRequestList(props) {
         initialValue={initialSearchBody.searchText}
         searchFunction={updateSearchText}
       />
-      {dataComponent}
+      <TableResultWrapper error={error} loading={loading} hasData={hasData}>
+        <TableComponent
+          pagination
+          data={formattedData}
+          handleTableRowClick={handleTableRowClick}
+          headRows={headRows}
+          page={searchBody.offset !== 0 ? searchBody.offset / searchBody.limit : 0}
+          setPage={updatePageNumber}
+          totalDoc={data.totalDoc}
+          rowsPerPage={searchBody.limit}
+          setRowsPerPage={updateRowsPerPage}
+          // Those are not displayed on small screens (mobile devices etc)
+          unprioritizedRows={['email', 'additionalInfo']}
+        />
+      </TableResultWrapper>
+
     </div>
   );
 }
-
-IsbnPublisherRequestList.propTypes = {
-  authenticationToken: PropTypes.string,
-  history: PropTypes.object.isRequired
-};
 
 export default IsbnPublisherRequestList;
